+++ plugins/tasks/ansible/src/main/java/com/walmartlabs/concord/plugins/ansible/RunPlaybookTask2.java
@@ -37,6 +37,8 @@ import java.nio.file.*;
 import java.nio.file.attribute.PosixFilePermission;
 import java.util.*;
 
+import static com.walmartlabs.concord.project.InternalConstants.Context.EVENT_CORRELATION_KEY;
+
 @Named("ansible2")
 public class RunPlaybookTask2 implements Task {
 
@@ -101,6 +103,10 @@ public class RunPlaybookTask2 implements Task {
         processStrategy(tmpDir);
 
         final Map<String, String> env = addExtraEnv(defaultEnv(workDir), args);
+        UUID eventCorrelationId = (UUID) context.getVariable(EVENT_CORRELATION_KEY);
+        if (eventCorrelationId != null) {
+            env.put("CONCORD_EVENT_CORRELATION_ID", eventCorrelationId.toString());
+        }
 
         GroupVarsProcessor groupVarsProcessor = new GroupVarsProcessor(secretReader);
         groupVarsProcessor.process(txId, args, workDir);